Fix: cleaning the code
